--- app/(analysis)/_components/channel-stats/index.ts
@@ -1,1 +1,0 @@
-export { ChannelStats } from "./channel-stats"
