--- app/(analysis)/_components/key-indicator/index.ts
@@ -1,1 +1,0 @@
-export { KeyIndicator } from "./key-indicator"
