+++ app/(auth)/layout.tsx
@@ -1,7 +1,13 @@
-const AuthLayout = ({ children }: { children: React.ReactNode }) => {
+import { CardWrapper } from "./_components/card-wrapper"
+
+export default function AuthLayout({
+  children,
+}: {
+  children: React.ReactNode
+}) {
   return (
-    <div className="flex h-screen items-center justify-center">{children}</div>
+    <div className="flex h-screen items-center justify-center">
+      <CardWrapper title="サインイン">{children}</CardWrapper>
+    </div>
   )
 }
-
-export default AuthLayout
