+++ app/(main)/comments/_components/badge-button/badge-button.tsx
@@ -4,10 +4,8 @@ import { useQueryState } from "next-usequerystate"
 
 import { cn } from "@/lib/utils"
 import { Badge, type BadgeProps } from "@/components/ui/badge"
-import {
-  isTabValue,
-  type TabValue,
-} from "@/app/(main)/comments/_components/types"
+
+import { isTabValue, type TabValue } from "../types"
 
 interface BadgeButtonProps extends BadgeProps {
   value: TabValue
