+++ app/(main)/comments/_components/index.ts
@@ -1,0 +1,3 @@
+export * from "./comments"
+export * from "./with-tab"
+export * from "./types"
