+++ app/(main)/layout.tsx
@@ -5,7 +5,7 @@ import { DEFAULT_LOGOUT_REDIRECT } from "@/routes"
 
 import { WithNavbar } from "@/components/layouts/with-navbar"
 
-async function MainLayout({ children }: PropsWithChildren) {
+export default async function MainLayout({ children }: PropsWithChildren) {
   const session = await auth()
   if (!session) {
     redirect(DEFAULT_LOGOUT_REDIRECT)
@@ -13,5 +13,3 @@ async function MainLayout({ children }: PropsWithChildren) {
 
   return <WithNavbar>{children}</WithNavbar>
 }
-
-export default MainLayout
